requirements.txt
python-dotenv
mysql-connector-python
Faker
requests